chore: clean up core
* chore: move dataframe dedicated method into core_df
